LibWeb: Split IdentifierStyleValue out of StyleValue.{h,cpp}
