+++ Userland/Libraries/LibWeb/CMakeLists.txt
@@ -82,6 +82,7 @@ set(SOURCES
     CSS/StyleValues/GridTrackPlacementStyleValue.cpp
     CSS/StyleValues/GridTrackPlacementShorthandStyleValue.cpp
     CSS/StyleValues/GridTrackSizeStyleValue.cpp
+    CSS/StyleValues/IdentifierStyleValue.cpp
     CSS/Supports.cpp
     CSS/SyntaxHighlighter/SyntaxHighlighter.cpp
     CSS/Time.cpp
