+++ Userland/Libraries/LibWeb/HTML/HTMLTableCaptionElement.cpp
@@ -5,6 +5,7 @@
  */
 
 #include <LibWeb/Bindings/Intrinsics.h>
+#include <LibWeb/CSS/StyleValues/IdentifierStyleValue.h>
 #include <LibWeb/HTML/HTMLTableCaptionElement.h>
 
 namespace Web::HTML {
